app/requirements.txt
flask
mysql-connector-python
marshmallow